requirements.txt
fastapi==0.115.3
uvicorn==0.32.0
python-dotenv==1.0.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
alembic==1.13.3
pytest==8.3.3
pytest-asyncio==0.24.0
PyJWT==2.9.0





